Some Deco and Exterior
Found a little house thing that might work for the exterior of the lodge (everything in the ExteriorLodge folder) just make sure the foundation and stuff goes below the snow so the door doesn't look wonky. Also added sound to the fireplace, a rug by the fireplace, and a stuffed bear in Room 1. Moved player start to center of first floor as well to line up with the exterior door positioning.